requirements.txt
faiss-cpu
PyMuPDF
google-generativeai
sentence-transformers
gradio
numpy
python-dotenv
